requirements.txt
asgiref==3.4.0
Django==3.2.4
django-crispy-forms==1.12.0
gunicorn==20.1.0
hashids==1.3.1
pytz==2020.1
sqlparse==0.3.1
typing-extensions==3.10.0.0
djangorestframework==3.12.4
whitenoise==5.3.0
django-heroku